requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
blobfile==3.0.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.33.1
distro==1.9.0
edge-tts==7.0.2
executing==2.2.0
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.5
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.2
gensim==4.3.3
git-lfs==1.6
google==3.0.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
googleapis-common-protos==1.67.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lxml==5.3.1
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.1
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multitasking==0.0.11
mypy==1.14.1
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.78.1
opencv-python==4.11.0.86
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.8.0
sgmllib3k==1.0.0
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
srt==3.5.3
sseclient-py==1.8.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
waitress==3.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.20.0
yfinance==0.2.52
zipp==3.21.0
zmq==0.0.0
